Weishiy 2 (#6)
* Purposed organisation of assignment.

This code is mostly method stubs and uninitialised
variables. I tried to fill out as little as possible
but still giving an idea of how it all fits together. Take note also of
how enums are used to store data.

The board design is my own sensibilities (option 2), where there are
stateful DoorTile s and PlayerTile s that reference back to the player.
The coordinates are still necessary, to find where they are stored in
board.

* I have added four classes to the game: Card (interface), Characters (enum), Houses (enum), and Weapons (enum). These are the fundamental elements of the game. The three enum classes have been inherited from Card.

* extended the program. I create a Board class to store all information about player's locations, grids and room location.
Estate class to identify the attributes for each estate, which is the type, the items and door's location.
I also create a Location class to store player's move on,
and the Player's class to store player's attributes, which the player should have the character, location, hand's cards, and it's states(lose or win).

* add some comments for game class and borad class

* Changed interface of RefutationController, changed objects it interacts with.

* Added additional comments to related classes.

* Created first refutation test, pruned environment.

Made framework for further refutation tests, should be fairly easy to
make a large number of cases. Pruned environment, meaning removing
unrelated private methods and unrelated classes: this mostly refers to
classes related to board.

* Working on implementing first test, need to adjust dependencies to provide more info.

* Changed test (again) to accomodate need for RefutationController to access the board.

The RefutationController needs to access the board to ask what estate the player guessing is in.

* Changed format of test, changed around dependencies of RefutationController.

* Created model state diagram for RefutationController

* Changed structure of dependencies to be more intuitive.

* Trying to update Player list in mock object to be in order.

* Fixed up merge, manually changing references and merging folders, renaming/small code changes.

* Further trying to integrate `main`, changing main and the refutation test.

* add some comments for game class and borad class

* Changed logging with RefutationControllerTest for cards to be predictable.

* Changed format of test function to connect cards to players.

* Got first test working.

* Added more tests, added check for single refute.

* Editted documentation.

* Editted documentation and style.

* Added test for correct guess.

* Moved objects to parent package, to better merge.

* Added more tests.

* some unfinished changes

* Basic UI

* Removed util package no longer used.

* Changed some documentation, and public methods of other classes to better match proposed organisation.

* Changed testing to incorporate new teleport. Also added new test.

* ...

* added locationsToEstateCard back which was missed during merged

* fixed an compile error caused by merge

* added the accidentially deleted class-diagram.ump back

---------

Co-authored-by: Jeremy Kanal-Scott <jkanalscott@gmail.com>
Co-authored-by: weishiy <118136175+weishiy@users.noreply.github.com>
Merged by Joseph